Move RuianTracer logic to abstract BuildingTracerModule
